pyproject.toml
[tool.poetry]
name = "hikari-bot-official"
version = "0.2.0"
description = "Nonebot2 HikariBot For QQ-Official,支持战舰世界水表查询"
authors = ["benx1n <shirakamikanade@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/benx1n/HikariBot-Official"
repository = "https://github.com/benx1n/HikariBot-Official"
keywords = ["nonebot", "nonebot2", "qqbot", "wows", "wws","bot","stats"]
packages = [
    { include = "hikari_bot_official", from = "./src/plugins/"}
]

[tool.poetry.dependencies]
python = "^3.8.0"
nonebot2 = { version = "^2.1.0", extras = ["fastapi","websockets","httpx"] }
nonebot-adapter-qq = "^1.1.0"
nonebot-plugin-apscheduler = "^0.2.0"
httpx = { version = ">=0.24.0", extras = ["http2"] }
Jinja2 = "^3.0.0"
beautifulsoup4 = "^4.11.1"
orjson = "^3.8.11"
hikari-core = "^1.0.0"
oss2 = "^2.18.3"

[tool.poetry.group.dev.dependencies]
nb-cli = "^0.6.0"
nonebot-plugin-gocqhttp = "^0.6.2"
black = {version = "^23.1a1", allow-prereleases = true}
isort = "^5.12.0"

[tool.nonebot]
plugins = ["nonebot_plugin_apscheduler"]
#plugins = ["nonebot_plugin_apscheduler","hikari_bot_official"]
plugin_dirs = ["src/plugins"]

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
# default = true

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 150
skip-string-normalization = true

[tool.ruff]
line-length = 150
respect-gitignore = true
ignore-init-module-imports = true
select = [
    'F',  # pyflakes
    'E',  # pycodestyle errors
    'W',  # pycodestyle warnings
    'I',  # isort
    'C',  # flake8-comprehensions
    'B',  # flake8-bugbear
    'Q',  # flake8-quotes
    'PL',  # pylint
    'RUF',  # Ruff-specific rules
]
ignore = [
    'E501',
    "RUF001",
    "RUF002",
    "RUF003",
    "PLR2004",
    "C901",
    "PLR0912",
    "PLR0911",
    "PLW0603"
]
flake8-quotes = {inline-quotes = 'single', multiline-quotes = 'double'}